add metodo login

+++ src/main/java/com/reis/financeiro/repository/RegistroRepository.java
@@ -1,12 +1,12 @@
 package com.reis.financeiro.repository;
 
 import com.reis.financeiro.entities.Registro;
-import com.reis.financeiro.entities.User;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RegistroRepository extends JpaRepository<Registro, Integer> {
+public interface RegistroRepository extends JpaRepository<Registro, Long> {
     List<Registro> findByUserId(Long userId);
